Remove server util
Because server util is not effective for clean code, just making longer
codes, so do I.
